+++ jannovar-core/src/main/java/de/charite/compbio/jannovar/annotation/VCFAnnotationData.java
@@ -5,7 +5,7 @@ import com.google.common.base.Joiner;
 import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableSortedSet;
 
-import de.charite.compbio.jannovar.annotation.AnnotationLocation.RankType;
+import de.charite.compbio.jannovar.annotation.SmallVariantAnnotationLocation.RankType;
 import de.charite.compbio.jannovar.hgvs.nts.change.NucleotideChange;
 import de.charite.compbio.jannovar.hgvs.protein.change.ProteinChange;
 import de.charite.compbio.jannovar.reference.GenomeVariant;
@@ -58,7 +58,7 @@ class VCFAnnotationData {
 	/** additional messages for the annotation */
 	public ImmutableSortedSet<AnnotationMessage> messages = ImmutableSortedSet.<AnnotationMessage> of();
 
-	public void setAnnoLoc(AnnotationLocation annoLoc) {
+	public void setAnnoLoc(SmallVariantAnnotationLocation annoLoc) {
 		if (annoLoc == null)
 			return;
 		if (annoLoc.getRankType() != RankType.UNDEFINED) {